Remove forward-declarations
