+++ unbound_script.h
@@ -4,7 +4,6 @@
 #include <stdint.h>
 
 #include "errors.h"
-#include "forward-declarations.h"
 
 #ifdef __cplusplus
 
@@ -15,6 +14,7 @@
 
 namespace v8 {
 class UnboundScript;
+class Isolate;
 }  // namespace v8
 
 struct m_unboundScript {
@@ -22,14 +22,18 @@ struct m_unboundScript {
 };
 
 typedef v8::ScriptCompiler::CachedData* ScriptCompilerCachedDataPtr;
+typedef v8::Isolate v8Isolate;
 
 extern "C" {
 #else
 
 typedef struct v8ScriptCompilerCachedData v8ScriptCompilerCachedData;
 typedef const v8ScriptCompilerCachedData* ScriptCompilerCachedDataPtr;
+typedef struct v8Isolate v8Isolate;
 
 #endif
+
+typedef v8Isolate* IsolatePtr;
 
 typedef struct m_ctx m_ctx;
 typedef m_ctx* ContextPtr;
